requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-babel
flask-restx
marshmallow
psycopg2-binary
PyMySQL
python-dotenv
flask-mail
ics
openpyxl
weasyprint
gunicorn
Flask-Testing
Flask-WTF
WTForms-SQLAlchemy
email_validator
Faker
pymysql